pyproject.toml
[tool.poetry]
name = "soar-lite"
version = "0.1.0"
description = "CLI‑first security orchestration & automation engine (SOAR‑Lite)"
authors = ["Ameen Abubacker <ameen@abubacker.dev>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.9"
requests = "^2.31"
PyYAML = "^6.0"
sqlmodel = ">=0.0.13"
apscheduler = "^3.10"
structlog = "^23.3"
pydantic = ">=2,<3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3"
sphinx = "^7.2"
coverage = "^7.2"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

